% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTTinShiny.R
\docType{package}
\name{CTTinShiny}
\alias{CTTinShiny}
\alias{CTTinShiny-package}
\title{Shiny Interface for the CTT Package}
\description{
A Shiny interface developed in close coordination with the CTT package, 
providing a GUI that guides the user through CTT analyses.
}
\author{
John T. Willse \email{jtwillse@uncg.edu}
}
