% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCTT.R
\name{shinyCTT}
\alias{shinyCTT}
\title{shinyCTT}
\usage{
shinyCTT()
}
\value{
No values are returned to R. After the Shiny interface launches
all values are returned within the Shiny interface itself.
}
\description{
Launches the Shiny interface for CTT
}
\details{
Launches a Shiny interface that was developed in conjunction with the CTT package.
The interface was designed to be intuitive and guide users through analyses. Some
options are hidden until data are loaded. Also, some options (distractor analysis)
are hidden unless raw data AND a key file are loaded.
}
\examples{
\dontrun{
shinyCTT()
}

}
